Move property retrieval functions to property package
